+++ External/sloan/sloan.h
@@ -1,12 +1,13 @@
 #ifndef SLOAN__
 #define SLOAN__
 
-int     gegra_ (int *, int *, int *, int *, int *, int *, int *, int *, int *);
-int     label_ (int *, int *, int *, int *, int *, int *, int *, int *);
+//int     gegra_ (int *, int *, int *, int *, int *, int *, int *, int *, int *);
+int     gegra_ (int *, int *, int *, long *, long *, long *, int *, int *, int *);
+int     label_ (int *, int *, int *, int *, long *, int *, int *, int *);
 int     iorele_ (int *, int *, int *, int *, int *, int *, int *, int *);
-int profi1_ (int *, int *, int *, int *, int *, int *, int *);
-int diamtr_ (int *, int *, int *, int *, int *, int *, int *, int *, int *, int *);
-int number_ (int *, int *, int *, int *, int *, int *, int *, int *, int *, int *);
+int profi1_ (int *, long *, int *, int *, int *, int *, int *);
+int diamtr_ (int *, int *, int *, int *, long *, int *, int *, int *, int *, int *);
+int number_ (int *, int *, int *, int *, int *, int *, int *, long *, int *, int *);
 int i_dim(int value_1, int value_2);
 
 #endif
